AIRLINE_MANUFACTURER/requirements.txt
flask
torch
torchvision
pillow
